sock/uring: Add the async network I/O support for socket
This patch is used to add the async network I/O support.
If code is configured with I/O uring, --with-uring, we  can
use io uring in Linux (version >=5.4-rc3). Otherwise, it will use
libaio.

PS: We also make VPP's default priority > uring, because
for the iSCSI or sock test linked with VPP. It tests VPP with
a given address (which is not a special VPP can only open address),
so using uring can also listen those address succefully. And if we make
uring with priority > VPP, actually, VPP will not tested in those cases.

Signed-off-by: Ziye Yang <ziye.yang@intel.com>
Change-Id: Iaaf43130bee5db4d0da1083d3e5428456ff2d2ba