Update patch set 4
Patch Set 4: Code-Review-1

Should lvol resize be considered in this context as well ?
Original intention of setting and keeping this property on creation, was to affect all future cluster releases.
This allowed user to define that the data written to particular lvol should always be overwritten with zeroes when no longer in use.
On the other hand, one can argue that the data can be overwritten by user while lvol bdev is still accessible.
A patch could be made that adds option for resize, same as delete in this series.

I'm not set on either way, just wanted to make sure all factors are considered.

Patch-set: 4
Reviewer: Gerrit User 1011221 <1011221@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1