Update patch set 6
Patch Set 6: Code-Review-1

This wasn't implemented internally, but the API does need to keep this on the create side. The behavior that can be configured is what happens when a cluster is released from the blob, which can happen dynamically during the lifetime of that blob in response to writes, resizes, or ultimately a delete.

The reason to clear the clusters when they're released is two-fold.

1) You may want to prevent "old" data from being visible in newly allocated blobs, if they are assigned a cluster that was previously used elsewhere. In this case, you want to make sure that blobs get zeroed out.
2) You likely want to trim clusters as they're released for performance reasons.

When you're just benchmarking, however, you may want to do neither of those since they can take up some time. So we wanted to make the behavior configurable.

Patch-set: 6
Reviewer: Gerrit User 1010531 <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1