Update patch set 22
Patch Set 22: Code-Review-1

It's unclear to me whether lvolstore needs to be zeroing/unmapping the data region on lvolstore create or delete at all. Couldn't that be left to an upper layer? If the user wants to do that before creating an lvolstore they are certainly free to do so.

For lvols it's more complicated. Certainly after an lvolstore is deleted the user can then go secure erase the device. But if just an lvol is deleted/resized/whatever, to avoid security issues with the cluster being reused later the user needs to be able to control the behavior here. The real configuration is what happens when a cluster is released - not when an lvol is deleted.

Patch-set: 22
Reviewer: Gerrit User 1010531 <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1