Update patch set 2
Patch Set 2:

(2 comments)

> Patch Set 2: Code-Review-1
> 
> Hi Jin,
> 
> Outstanding bdev IOs are managed by bdev modules and so io_submitted is duplicated.
> 
> I think it may be possible for us to pass io channel to bdev module and get the corresponding bdev IOs.
> 
> So sorry for back and force but I want to mark -1 and know any comment.
> 
> Thanks
> Shuhei

Hi shuhei,

sorry for the late response. This patch is for checking the overtime bdev_io in a specific bdev_channel(you can see the second patch that we supply an API when the user create io_channel with this API can supply a timeout callback and timeout limit). So I add a list in the bdev_channel to count all the bdev_ios that were submitted to this channel.

Patch-set: 2
Reviewer: Gerrit User 1015333 <1015333@d5d70762-12d0-45a1-890d-524b12d3f735>