nvme_ctrlr: enable the admin qpair at init start.
This function has historically waited until we have to do a listen
before enabling the admin qpair. That is a very PCIe-centric mindset.
For fabric controllers, a lot ofthe early initialization operations such
as get_cc and set_cc are handled through the admin qpair so it should be
enabled before those operations.

Change-Id: Icd162657d01a85c227a3f20c295d0208e07ce44d
Signed-off-by: Seth Howell <seth.howell@intel.com>