Update patch set 8
Patch Set 8:

> Patch Set 8: Verified-1
> 
> Build failed. Results https://dqtibwqq6s6ux.cloudfront.net/public_build/autotest-per-patch_46969.html

The nvmf_phy_autotest failed on this patch. I was worried that it was a race condition in the new code causing a latent failure, but I could not find any possible sources of the race. Upon closer inspection of the logs, it looks like I killed the target application before the controller was able to fully initialize the controllers. This is what caused the failure. The solution is to delay until the probe and attach callbacks can be called before killing the target.
To summarize, this failure was caused by the test structure, and not by any races in the code.

Patch-set: 8
Reviewer: Gerrit User 1011223 <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>