nvmf/tcp: fix the state machine issue if data is already read
Since we use big buffer to read the data, so the incoming
data may already be read. If we do not add this,
we will never hand the payload of the incapsule data.

Since the data is already read into the temp big buffer,
and there will be no read event will be generated again.

Signed-off-by: Ziye Yang <ziye.yang@intel.com>
Change-Id: Ib195d57cc2969235203c34664115c3322d1c9eae