Update patch set 2
Patch Set 2:

incapsule data requests for data cannot be always allocated. It still needs to waits the buffer, if this request is not the first req which waits for the buffer. It can be occur, if there are already pressure tests which compete for buffers.

Patch-set: 2
Reviewer: Gerrit User 1011275 <1011275@d5d70762-12d0-45a1-890d-524b12d3f735>