Update patch set 13
Patch Set 13:

Hi Guys.  Sorry to do this to you, the new tip of the chain is here: https://review.gerrithub.io/c/spdk/spdk/+/472209

and it works as documented before we started all of this.

lvol_store: accepts the option on create and if only affects how the data space is initialed on create
lvol: accepts the option on create and stores it in blob metadata for later use on resize/delete just like it used to but it works now (or it should).

NOTE: when an lvol or blob is opened it's possible to pass open options that include clear_method. For create/delete these are internal. All internal clear method options init to DEFAULT (the enum value) which will be lumped together with UNMAP (the enum) to do unmap.  

If for some reason the stored clear_method is different than the one passed in, the one passed in will be used unless its DEFAULT then the stored will be used.  It's a little confusing and I'm open to suggestions but other than adding a new enum value of UNSPECIFIED, I chose to take "DEFAULT", when passed in, to sorta mean "do what makes sense" which means use the stored value.  If the passed in value is clearly something non-default like WZ, it will override the stored clear_method option. You shouldn't have to worry about this internal detail, just wanted to let you know.

Patch-set: 13
Reviewer: Gerrit User 1011289 <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>