Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > What's a reason for removing raid_fn_table ?
> 
> Hi,
> 
> I'm re-introducing a different function table later in the series.

would you consider to design a generic interface, which is not only used for raid, but also for other "one device map to other device(s)" modules? I submitted a linear module patch set, but there are lots of duplicate code with raid:
https://review.gerrithub.io/c/spdk/spdk/+/472103
In fact, more than 90% code in my patch set was copied from the current raid module. So let the raid code be reusable for linear module would be a good idea. If the raid module doesn't assume that "strip_size" is a mandatory parameter, it should be good enough to let linear module reuse it.

Patch-set: 1
CC: Gerrit User 1017170 <1017170@d5d70762-12d0-45a1-890d-524b12d3f735>