opal: Add a state machine to avoid invalid operations
If an Opal drive is enabled and someone else still tries
to init the drive, several times later, the drive will be
locked out. If the Opal drive has already been reverted,
and if revert is triggered several times again, it also
will be locked out. Once locked out, a system reboot is
required.

Thus, adding a state machine to avoid this type of failure.
Not apply to the locked out situation due to wrong password
input for too many times.

Change-Id: I92f5cb26228bb883af97af209c29b19c564473d1
Signed-off-by: Chunyang Hui <chunyang.hui@intel.com>