scripts/ceph: unmount the right dev.
According to my check, I found that we should
umount /dev/loop200p2. The name of the dev used in the
start.sh and stop.sh is not same.

Then there an issue reported in the script:

umount /dev/loop200
umount: /dev/loop200: not mounted.
I think that this is an issue happens in our every
test. And it may influence the further test.

Signed-off-by: Ziye Yang <ziye.yang@intel.com>
Change-Id: I10c460b4e7760f321fd1a1cd87caa69d2911037b