Update patch set 10
Patch Set 10:

> Patch Set 10: Code-Review+1
> 
> Guess I should go read a spec or something but real quick, what's the purpose for SPDK_NVMF_TCP_RECV_BUF_SIZE_FACTOR (set to 4) as used in these calculations? Thanks :)

That's just how big we want the recv buffer to be in units of whatever the in-capsule data size is negotiated to be. In this case, we have decided we want it to be at least 4x the size of the in-capsule data. I'm not really sure why it's specified that way.

Patch-set: 10
Reviewer: Gerrit User 1010531 <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>