Update patch set 2
Patch Set 2: Code-Review-1

So one of the whole ideas here with this example was to have cli_context be allocated in main and passed to calls and via callbacks thus no use of globals through all the async operations. 

Plenty of our other examples use globals but this one doesn't so I'd rather not make cli_context a global here if we don't have to (and even if we did it needs to be named g_cli_context).

Is there even a reason why the context needs to be set blobcli_shutdown()? Isn't that the last function executed by the framework?  If I take your patch and remove the setting of context from the shutdown routine it seems to work fine.

Maybe provide the exact steps you ran to get into the situation that you felt like you needed to change this?  Thanks!!

Patch-set: 2
Reviewer: Gerrit User 1011289 <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1