bdev: Add is_outstanding flag to bdev_io to complete unsubmitted I/O
When spdk_bdev_zcopy_start() is used or unaligned I/O is split,
spdk_bdev_io_get_buf() is called before I/O is submitted to bdev
module, and then if buffer pool is empty, I/O is queued.

At this time, if I/O is aborted, it is dequeued and spdk_bdev_io_complete()
is called to it.  However it is not accounted in outstanding I/Os yet,
and spdk_bdev_io_complete() will not work correctly.

This patch adds io_outstanding flag to struct bdev_io and use it
in _spdk_bdev_abort_buf_io() by using _spdk_bdev_abort_queued_io()
as a reference.

If we use spdk_bdev_zcopy_start/end(), we will need this patch.

Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Change-Id: I2a876aa6853b017b8c799b7518cc6ee428037c1a