iscsi: Fix wrong fix to close the being hot-removed LUN
About the specific commit Iad6ecdc37493fa9f2d7ccab262a2c75dac2fcd48
the patch didn't implement the change described in the commit message.

If the connection is in exited state, LUN is not ensured to be
closed yet. So this patch changes the condition to check if the
target LUN is closed or not directly.

Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>
Change-Id: Ic57646f3bed80589e9025109e029088b1fe5ed9b