Update patch set 3
Patch Set 3:

Hi Artur,

Related with Paul's comment, you are changing to use not bdev_io but raid_bdev_io throughout, and that is so nice idea, and remaining patches are really good too.

On the other hand, unit tests may not be unified to use raid_bdev_io, raid_bdev_io_channel, and other raid_bdev_io_*.

If possible, will you check unit test if we can do the same improvement for it too?

-Shuhei

Patch-set: 3