bdev/opal: Add state check for rpc commands
In some cases, we can jump out earlier and to prevent
unexpected beviour. So add state check for all the rpc
commands.

All rpc commands should check if it is busy.
Besides, in some cases, sending specific rpc commands
should also check if it is not in specific state. For example,
if the opal_dev is in Default state, the revert cmd should
return immediately.

Change-Id: If7e0e99ad82936c997db40ecd3592f68a10f4577
Signed-off-by: Chunyang Hui <chunyang.hui@intel.com>