pci: fix the hotplug issue
DPDK intr thread is designed that it can't free the src callback
in the intr thread. So I think we can't detach the PCI device in
the intr thread as it needs to unregister the VFIO notification
which will be not successful but it still can free the device.
So at the next req notification in the handler function, we meet
the freed device.

Fix #994

Change-Id: Id4b45a2d0fe6b45b132355d59471bc80240fad70
Signed-off-by: Jin Yu <jin.yu@intel.com>