bdev_merge:
- Add asynchronous write to merge_bdev
- Add double or multiple buffers to hold data. While using multiple buffers, as the buffer count adds,
the ratio of read to write will become smaller, which means it can adjust the ratio of read to write by
changing the buffers count. The largest count is 31. However, to reach the best ratio of read to write,
it is contingent on the physical devices used.
- Notice: Before the merge bdev closes, it should wait for all I/Os in the queue(to be finished).
So currently, when put the I/O channel. it will find that some outstanding I/Os still remain
not finished or in the queue and may abort them, which can be seen as errors, and the merge bdev
may not be able to handle these I/O at all.
- Fix: currently using sleep() to wait for finishing these outstanding I/Os.

Signed-off-by: Liu.Lanfeng <lanfeng.liu@intel.com>
Change-Id: If47761dfd8f9cc30ef8eb77fe355de96f647e2da