bdev_merge:
- Add asynchronous write to merge_bdev
- Add multiple buffers for holding data. Currently we can adjust the ratio of read to write by
changing the count of buffers, and usually the more the buffer, the higher the ratio. However,
to reach a best performance, it is contingent to the physical devices used.
- Notice: Before the merge bdev closes, it will wait for all the I/O in the queue(to be
finished). However if the frontend closes before the backend and checks the I/Os submitted
previously(such as fio), it will find that some I/Os still remain not finished or in the queue
and may abort them, which can be seen as errors, and the merge bdev may not be able to
handle these I/O at all.
- Fix: currently using sleep() to wait for finishing these outstanding I/Os.

Signed-off-by: Liu.Lanfeng <lanfeng.liu@intel.com>
Change-Id: If47761dfd8f9cc30ef8eb77fe355de96f647e2da

bdev_merge: Add double or multiple buffers to hold data

While using multiple buffers, as the buffer count adds, the ratio of read to write will become smaller,
which means it can adjust the ratio of read to write by changing the buffers count. The largest count
is 32. However, to reach the best ratio of read to write, it is contingent on the physical devices used.

Change-Id: I576194500c1bdb9f55944f933c60ad8ded68b0b6
Signed-off-by: Liu.Lanfeng <lanfeng.liu@intel.com>